# remote/read_handler: pool input to Marshal()