Use a sync.Pool to reuse byte slices between calls to Marshal() in the remote read handler.

Fixes #11232.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>