@GiedriusS 2022-09-27
Tried out with https://github.com/thanos-io/thanosbench that with this change I am able to retrieve ~33MiB in 5 seconds from Prometheus + Sidecar. Without this change, I am able to retrieve only ~31MiB in 5 seconds. So, only a small improvement but still an improvement nonetheless.