@GiedriusS 2022-11-15
Added a small microbenchmark. Here is the diff:

```
name                   old time/op    new time/op    delta
StreamReadEndpoint-16    43.7µs ± 3%    39.9µs ± 3%  -8.70%  (p=0.004 n=5+6)

name                   old alloc/op   new alloc/op   delta
StreamReadEndpoint-16    28.8kB ± 0%    26.9kB ± 0%  -6.60%  (p=0.000 n=5+4)

name                   old allocs/op  new allocs/op  delta
StreamReadEndpoint-16       292 ± 0%       277 ± 0%  -5.14%  (p=0.002 n=6+6)
```